+++ packages/core/mashroom-json-schemas/type-definitions/mashroom-server-config.d.ts
@@ -31,6 +31,10 @@ export interface MashroomServerConfig {
    */
   enableHttp2?: boolean;
   /**
+   * Configure if reverse proxies should be trusted (e.g., with secure cookies). See https://expressjs.com/en/guide/behind-proxies.html for details (Default: 1)
+   */
+  trustProxy?: boolean | number | string;
+  /**
    * The start page
    */
   indexPage?: string;