+++ packages/core/mashroom/type-definitions/api.ts
@@ -334,6 +334,7 @@ export type MashroomServerConfig = {
     readonly httpsPort: number | null | undefined;
     readonly tlsOptions: TlsOptions | null | undefined;
     readonly enableHttp2: boolean;
+    readonly trustProxy: boolean | number | string;
     readonly xPowerByHeader: string | null | undefined;
     readonly serverRootFolder: string;
     readonly tmpFolder: string;